Revise data model, refactor TH -> generic instance
Bumps new API version

We revise data model to allow the client/user to specify the action that
they would like to do. As a result some refactoring in the ToSchema instances
was made, specifically to allow enumerations with nested content.

On the same token we refactor out the TH derivation of the ToSchema (Key a)
things into ToSchema generic instances that match on all Key-s (that has
a proper ToSchema for it's backend).